+++ src/node/neighborhood/transaction_response.go
@@ -3,5 +3,5 @@ package neighborhood
 type TransactionResponse struct {
 	SenderAddress    string
 	RecipientAddress string
-	Value            float32
+	Value            uint64
 }
